add the database folder
