--- server/model/user.ts
@@ -1,32 +1,0 @@
-import { Model, DataTypes } from 'sequelize';
-import sequelize from './util/db';
-
-class User extends Model { }
-
-User.init({
-  id: {
-    type: DataTypes.INTEGER,
-    primaryKey: true,
-    autoIncrement: true
-  },
-  username: {
-    type: DataTypes.TEXT,
-    allowNull: false
-  },
-  // Do we even store the hashed password in the database?
-  passwordHash: {
-    type: DataTypes.TEXT,
-    allowNull: false
-  },
-  email: {
-    type: DataTypes.TEXT,
-    allowNull: false
-  }
-}, {
-  sequelize,
-  underscored: true,
-  timestamps: false,
-  modelName: 'note'
-})
-
-export { User };
